backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlmodel==0.0.22
alembic==1.14.0
psycopg2-binary==2.9.10
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
httpx<0.28.0
google-generativeai
openai
